requirements.txt
autopep8==1.3.5
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==2.0.4
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.2
pycodestyle==2.4.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
tqdm==4.23.0
uritemplate==3.0.0
urllib3==1.22
